requirements/base.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
dj-rest-auth==7.0.1
Django==4.2.7
django-cors-headers==4.3.1
django-extensions==4.1
django-filter==25.1
django-guardian==3.2.0
django-model-utils==5.0.0
django-ratelimit==4.1.0
django-redis==6.0.0
django-rest-auth==0.9.5
django-taggit==6.1.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
factory_boy==3.3.3
Faker==37.8.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg==3.2.10
psycopg-binary==3.2.10
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
requests==2.31.0
rpds-py==0.27.1
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
